Use updated license text from https://spdx.org/licenses/AMPAS.html as it has explicit and refined statements about distribution and is the license referenced in the new SPDX headers.
Principal differences around the concept of "distribution" are as follows:

*aces-dev* License

A worldwide, royalty-free, non-exclusive right to copy, modify, create derivatives, and use, in source and binary forms, is hereby granted, subject to acceptance of this license.

[...]

Copies of source code, in whole or in part, must retain the above copyright notice, this list of conditions and the Disclaimer of Warranty.

*AMPAS* License

A world-wide, royalty-free, non-exclusive right to distribute, copy, modify, create derivatives, and use, in source and binary forms, is hereby granted, subject to acceptance of this license.

[...]

Redistributions of source code must retain the above copyright notice, this list of conditions and the Disclaimer of Warranty.